+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/mailer_with_pgp_signer_without_secret_key.php
@@ -1,0 +1,10 @@
+<?php
+
+$container->loadFromExtension('framework', [
+    'mailer' => [
+        'dsn' => 'smtp://example.com',
+        'pgp_signer' => [
+            'enabled' => true,
+        ],
+    ],
+]);