+++ src/Symfony/Component/Mime/Part/PgpEncryptedMessagePart.php
@@ -1,0 +1,35 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Mime\Part;
+
+/**
+ * @author Florent Morselli <florent.morselli@spomky-labs.com>
+ *
+ * Original idea by PuLLi <the@pulli.dev>
+ *
+ * @internal
+ */
+final class PgpEncryptedMessagePart extends AbstractPgpPart
+{
+    public function __construct(string $body)
+    {
+        parent::__construct($body);
+        $this->getHeaders()->addParameterizedHeader('Content-Disposition', 'inline', [
+            'filename' => 'msg.asc',
+        ]);
+    }
+
+    public function getMediaSubtype(): string
+    {
+        return 'octet-stream';
+    }
+}